+++ src/core/themes/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Barrel für das Theme-Modul.
+ *
+ * Verwendung:
+ *   import { getThemeOrFallback, themeToCssVars } from "@/core/themes";
+ */
+
+export * from "./registry";
+export * from "./theme-resolver";